add support for min, max, minu and maxu instructions from Zba

+++ include/fenn/ise/riscv.h
@@ -46,7 +46,7 @@ private:
 class ISE_EXPORT ScalarDataMemory
 {
 public:
-    ScalarDataMemory(size_t numBytes, size_t startAddressBytes, uint8_t poissonVal);
+    ScalarDataMemory(size_t numBytes, uint32_t startAddressBytes, uint8_t poissonVal);
 
     uint32_t read32(uint32_t addr) const;
     void write32(uint32_t addr, uint32_t value);
@@ -65,7 +65,7 @@ public:
     size_t getSizeBytes() const{ return m_Data.size(); }
 
 private:
-    size_t m_StartAddressBytes;
+    uint32_t m_StartAddressBytes;
     std::vector<uint8_t> m_Data;
 };
